+++ examples/test_cookbook.py
@@ -148,7 +148,7 @@ class sample_class:
 			{
 				'p': [0], # "path" to argument containing the pointer
 				'l': [1], # "path" to argument containing the length
-				'_t': ctypes.c_double, # type of argument (optional, default char/byte): sizeof(type) * length == bytes
+				't': 'c_double', # type of argument (optional, default char/byte): sizeof(type) * length == bytes
 				'_c': DoubleArray # custom datatype
 				}
 			]
@@ -161,7 +161,7 @@ class sample_class:
 			{
 				'p': [0], # "path" to argument containing the pointer
 				'l': [1], # "path" to argument containing the length
-				'_t': ctypes.c_float # type of argument (optional, default char/byte): sizeof(type) * length == bytes
+				't': 'c_float' # type of argument (optional, default char/byte): sizeof(type) * length == bytes
 				}
 			]
 		self.__bubblesort__.argtypes = (ctypes.POINTER(ctypes.c_float), ctypes.c_int)