+++ src/zugbruecke/core/routine_server.py
@@ -114,7 +114,7 @@ class routine_server_class():
 		try:
 
 			# Pack memory for return
-			self.data.server_pack_memory_list(arg_memory_list, self.memsync_d)
+			self.data.server_pack_memory_list(args_list, arg_memory_list, self.memsync_d)
 
 			# Get new arg message list
 			arg_message_list = self.data.arg_list_pack(args_list, self.argtypes_d)
@@ -134,7 +134,7 @@ class routine_server_class():
 				'exception': None
 				}
 
-		except:
+		except Exception as e:
 
 			# Push traceback to log
 			self.log.err(traceback.format_exc())
@@ -151,7 +151,7 @@ class routine_server_class():
 		self.restype_d = restype_d
 
 		# Store memory sync instructions
-		self.memsync_d = memsync_d
+		self.memsync_d = self.data.unpack_definition_memsync(memsync_d)
 
 		try: